refactor: migrate screen-local data providers to ViewModel files (closes #50)
Migrated 9 data-fetching providers from screen files into dedicated *_view_model.dart files:

robot_capabilities/:
- components_screen.dart → components_view_model.dart (componentsProvider)
- contribute_history_view.dart → contribute_history_view_model.dart (contributeHistoryProvider)
- capabilities_widgets.dart: removed unused imports (CommandScope, robotRepositoryProvider)
- research_screen.dart: add import for personal_research_view_model

fleet_leaderboard/:
- fleet_leaderboard_screen.dart → fleet_leaderboard_view_model.dart (researchStatusProvider)
- personal_research_card.dart → personal_research_view_model.dart
  (userRrnProvider, personalResearchProvider, personalResearchRrnProvider)

robot_detail/:
- robot_detail_screen.dart → robot_detail_view_model.dart (latestVersionProvider)
- orchestrator_screen.dart → orchestrator_view_model.dart (orchestratorsProvider)

alerts/:
- alerts_screen.dart → alerts_view_model.dart (alertsProvider)

Also adds shared/loading_view.dart + shared/shared.dart (full-screen loading widget)
used by the refactored screens.
