+++ lib/ui/shared/loading_view.dart
@@ -1,0 +1,26 @@
+/// Full-screen loading state widget.
+library;
+
+import 'package:flutter/material.dart';
+
+class LoadingView extends StatelessWidget {
+  final String? message;
+
+  const LoadingView({super.key, this.message});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const CircularProgressIndicator(),
+          if (message != null) ...[
+            const SizedBox(height: 16),
+            Text(message!, style: Theme.of(context).textTheme.bodySmall),
+          ],
+        ],
+      ),
+    );
+  }
+}
