+++ lib/ui/shared/shared.dart
@@ -1,0 +1,13 @@
+/// Barrel export for shared UI widgets.
+library;
+
+export 'adaptive_navigation.dart';
+export 'confirmation_dialog.dart';
+export 'empty_view.dart';
+export 'error_view.dart';
+export 'google_sign_in_button.dart';
+export 'loading_skeleton.dart';
+export 'loading_view.dart';
+export 'pipeline_explainer.dart';
+export 'robot_card.dart';
+export 'status_indicator.dart';
